+++ lib/karafka/pro/routing/features/virtual_partitions/topic.rb
@@ -26,18 +26,24 @@ module Karafka
             # @param offset_metadata_strategy [Symbol] how we should match the metadata for the
             #   offset. `:exact` will match the offset matching metadata and `:current` will select
             #   the most recently reported metadata
+            # @param reducer [nil, #call] reducer for VPs key. It allows for using a custom
+            #   reducer to achieve enhanced parallelization when the default reducer is not enough.
             # @return [VirtualPartitions] method that allows to set the virtual partitions details
             #   during the routing configuration and then allows to retrieve it
             def virtual_partitions(
               max_partitions: Karafka::App.config.concurrency,
               partitioner: nil,
-              offset_metadata_strategy: :current
+              offset_metadata_strategy: :current,
+              reducer: nil
             )
               @virtual_partitions ||= Config.new(
                 active: !partitioner.nil?,
                 max_partitions: max_partitions,
                 partitioner: partitioner,
-                offset_metadata_strategy: offset_metadata_strategy
+                offset_metadata_strategy: offset_metadata_strategy,
+                # If no reducer provided, we use this one. It just runs a modulo on the sum of
+                # a stringified version, providing fairly good distribution.
+                reducer: reducer || ->(virtual_key) { virtual_key.to_s.sum % max_partitions }
               )
             end